backend/requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
telethon==1.28.1
solders==0.16.1
solana==0.29.0
python-dotenv==1.0.0
